Adds edible resources
Makes it so that some resources are edible and restore player stats. Fixes the bug that didn't allow more than one building part in the UI. Made it so that building parts can snap onto the grid when being placed. Adds a berry bush that can be gathered and eaten.